+++ integration_tests/suite/test_documentation.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
@@ -19,6 +19,6 @@ class TestDocumentation(BaseTestCase):
 
     def test_documentation_errors(self):
         port = self.service_port(9497, 'auth')
-        api_url = 'https://localhost:{port}/0.1/api/api.yml'.format(port=port)
-        api = requests.get(api_url, verify=False)
+        api_url = 'http://localhost:{port}/0.1/api/api.yml'.format(port=port)
+        api = requests.get(api_url)
         validate_v2_spec(yaml.safe_load(api.text))